01-foundations/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing and validation
pydantic==2.5.0
python-dotenv==1.0.0

# Vector embeddings and search
sentence-transformers>=2.3.0
faiss-cpu==1.7.4
numpy==1.24.3

# LLM integration
openai==1.3.7
anthropic==0.7.8

# Text processing
langchain
langchain-community
langchain-text-splitters

# Utilities
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
